+++ src/paredit/statusbar.ts
@@ -5,68 +5,63 @@ import * as paredit from './extension';
 
 export class StatusBar {
 
-    private _enabled: Boolean;
     private _visible: Boolean;
-    private _keyMap: String;
+    private _keyMapConfig: paredit.KeyMapConfig;
 
     private _toggleBarItem: StatusBarItem;
 
-    constructor(keymap: String) {
+    constructor(keyMapConfig: paredit.KeyMapConfig) {
         this._toggleBarItem = window.createStatusBarItem(StatusBarAlignment.Right);
         this._toggleBarItem.text = "(λ)";
         this._toggleBarItem.tooltip = "";
         this._toggleBarItem.command = 'paredit.togglemode';
-        this._enabled = false;
         this._visible = false;
-        this.keyMap = keymap;
+        this.keyMapConfig = keyMapConfig;
 
-        paredit.onPareditKeyMapChanged((keymap) => {
-            this.keyMap = keymap;
-        }) 
+        paredit.onPareditKeyMapChanged((keyMapConfig) => {
+            this.keyMapConfig = keyMapConfig;
+        })
     }
 
-    get keyMap() {
-        return this._keyMap;
-    }
+    updateUIState() {
+        let keyMap = this._keyMapConfig.keyMap.trim().toLowerCase();
+        let keybindingsEnabled = this._keyMapConfig.keybindingsEnabled;
 
-    set keyMap(keymap: String) {
-        
-        switch (keymap.trim().toLowerCase()) {
-            case 'original':
-                this._keyMap = 'original';
-                this.enabled = true;
-                this.visible = true;
-                this._toggleBarItem.text = "(λ)";
-                this._toggleBarItem.tooltip = "Toggle to strict Mode"
-                break;
-            case 'strict':
-                this._keyMap = 'strict';
-                this.enabled = true;
-                this.visible = true;
-                this._toggleBarItem.text = "[λ]";
-                this._toggleBarItem.tooltip = "Toggle to original Mode"
-                break;
-            default:
-                this._keyMap = 'none';
-                this.enabled = false;
-                this.visible = true;
-                this._toggleBarItem.text = "λ";
-                this._toggleBarItem.tooltip = "Calva Paredit Keymap is set to none, Toggle to Strict Mode is Disabled"
-        }
-    }
-
-    get enabled() {
-        return this._enabled;
-    }
-
-    set enabled(value: Boolean) {
-        this._enabled = value;
-
-        if (this._enabled) {
-            this._toggleBarItem.color = undefined;
+        if (keybindingsEnabled) {
+            switch (keyMap) {
+                case 'original':
+                    this.visible = true;
+                    this._toggleBarItem.text = "(λ)";
+                    this._toggleBarItem.tooltip = "Toggle to strict Mode";
+                    this._toggleBarItem.color = undefined;
+                    break;
+                case 'strict':
+                    this.visible = true;
+                    this._toggleBarItem.text = "[λ]";
+                    this._toggleBarItem.tooltip = "Toggle to original Mode";
+                    this._toggleBarItem.color = undefined;
+                    break;
+                default:
+                    this.visible = true;
+                    this._toggleBarItem.text = "λ";
+                    this._toggleBarItem.tooltip = "Calva Paredit Keymap is set to none, Toggle to Strict Mode is Disabled"
+                    this._toggleBarItem.color = statusbar.color.inactive;
+            }
         } else {
+            this.visible = true;
+            this._toggleBarItem.text = "λ";
+            this._toggleBarItem.tooltip = "Calva Paredit Keybindings Enabled is set to false"
             this._toggleBarItem.color = statusbar.color.inactive;
         }
+    }
+
+    get keyMapConfig() {
+        return this._keyMapConfig;
+    }
+
+    set keyMapConfig(keyMapConfig: paredit.KeyMapConfig) {
+        this._keyMapConfig = keyMapConfig;
+        this.updateUIState();
     }
 
     get visible(): Boolean {